Refactor: update CONSTANT_KAFKA import paths and create new constants
