+++ src/modules/common/constants/constantsKafka.ts
@@ -2,6 +2,7 @@ export const CONSTANT_KAFKA = {
   SOURCE: {
     USER_SERVICE: 'UserService',
     PRODUCT_SERVICE: 'ProductService',
+    NOTIFICATION_SERVICE: 'NotificationService',
   },
 
   TOPIC: {
@@ -11,6 +12,9 @@ export const CONSTANT_KAFKA = {
     },
     PRODUCT: {
       CREATED: 'product_created',
+    },
+    NOTIFICATION: {
+      EMAIL: 'email_notification',
     },
   },
 };
